+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 
+## 0.0.3
+- resolve #10 : user color setting
+
 ## 0.0.2
-- resolve #2
-- resolve #3
+- resolve #2 : overviewRulerColor
+- resolve #3 : search bug fix
 
 ## 0.0.1